CP-30166: Recognise multi-queue blkback kthreads for QoS setting
For VBDs that use blkback as backend driver (e.g. for raw LVM VDIs), blkback now has a new way of recording the kthread-pid in xenstore, as there can now be multiple kthreads per VBD, each serving a separate queue. If a QoS policy is set for such a VBD, then xenopsd needs to apply it to all kthreads using the `ionice` command.

The old xenstore path for the kthread-pid is:

    /local/domain/0/backend/vbd/<domid>/<devid>/kthread-pid

In the new case, there may be multiple:

    /local/domain/0/backend/vbd/<domid>/<devid>/queue-0/kthread-pid
    /local/domain/0/backend/vbd/<domid>/<devid>/queue-1/kthread-pid
    (...)

Xenopsd now recognises both formats.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>