Fix cross-pool migration for qemu-trad VMs
Since 4e669d02, the `profile_of` function that was used unilaterally in
`set_internal_state` raises an exception if the device_model key in the
platform map is `qemu-trad`, which may be the case in cross-pool migrations.
For this reason, this function cannot be used.

However, we still need to set the profile field in the state, in the incoming
VM did not have one yet. In this case, we always assume a Qemu_trad profile.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>