CA-313265: Don't limit QEMU's file size when using file serial
When QEMU's emulated serial console is connected to a file, don't limit
the maximum file size to avoid the output being unexpectedly cropped.

An alternative fix which still maintains the file size limit would be to
have a separate daemon to which QEMU pipes the output. Since using file
serial is probably infrequently used, this fix should be fine for now.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>