CA-314170: wait for frontend_closed with qemu-dp too
When qemu-dp is used (SMAPIv3 SRs) then unplug_watch returned a dummy
watch that "fired" immediately.
This means that on a vbd unplug we went ahead and wiped the xenstore
keys for the disk while it was still in state 5 (Closing), before it has
reached state 6 (Closed).

This confused both qemu-dp and the guest which didn't know anymore what
state the disk was in and put/tried to process random garbage into/from the PV ring.

On a clean VBD unplug wait for the frontend to be closed even with
qemu-dp:
* if the key is already deleted then Watch.any_of will return
immediately (this is the case when cancelable_watch is called for the 2nd time)
* otherwise try to wait until the frontend reaches the correct state

If we update qemu-dp to version 4.x we may not need to do this anymore.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>