CP-30136: Fixed PCI slot assignment for passthrough devices in uefi device model
The assignment should be handled by the device-model.
For qemu-upstream-compat we let qemu decide (but we no longer use xl pci-attach)
For qemu-upstream-uefi we use 8+passthrough device index according to
the design:
         0: i440FX
         1: PIIX3
         2: VGA or empty
         3: Xen platform or empty
         4 - 5: NIC (limited to first 2)
         6: Xen PV
         7: NVME
         8+: vGPU and other pass-through devices

Devices 0-7 already got implemented as part of the initial UEFI work,
        and 8+ is implemented here.

In the future we would want XAPI to be in charge of this assignment
though.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>