XSI-321: Do FLR before de-assigning device.
Change the order of FLR and de-assigning. Previously we did de-assigning
before resetting a device, then if the device had a ongoing DMA, the
device may cause the memory corrumption when it was back to Dom0.

Resetting a device will notify the device to stop DMA.

Backport of dededb3c97b50d4164bc5596c11462445fe44fa6

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>