CA-318579 write serial device to xenstore for HVM
From XSI-353, since the upgrade to QEMU upstream the ability to access
the serial console of an HVM guest using xl console or xenconsole has
been broken.

This commit adds code to query QEMU for the path of the serial device
and writing it to xenstore:

  /local/domain/<domid>/serial/0/tty <ptypath>

Backport of 433805448ed3b674a441fb3d67f8471956c87793. This depends on
backported changes in ocaml-qmp.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>