CA-320215: use xl pci attach for PV guests
The new code in xenopsd only supports HVM guests.
It is unclear whether anyone would still be using PCI passthrough to PV guests,
we haven't explicitly documented that it is not supported,
so add it back and fall back to the old code for it.

We can deprecate it and remove it in the next version.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>