REQ-627 introduce type PCI.device'
This commit introduces a record type for a previously anonymous tuple
that is used to pass a mapping between a host PCI address and a guest
PCI address tp PCI.add.

The change is anticipation of adding a flag to the new record in the
context of NVIDIA SR-IOV GPUs.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>