REQ-627 don't advertise NVIDIA SR-IOV PCI devices
Plugging a PCI devices involves advertising the device to QEMU using QMP
command Device_add. This must not be done for NVDIA SR-IOV devices.

Supporting this requirement requires a bit of plumbing. The PCI.plug
function receives now a flag that indicates whether or not to advertise
the device using Device_add.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>