CA-320079: Add support for NVME namespaces
NVME namespaces allow having multiple disks hanging off the same NVME
controller. This requires some adjustment to the options passed to QEMU.

As an example, the options for a VM with two namesapces look like this:
  -device nvme,serial=nvme0,id=nvme0,addr=7
  -drive id=disk0,if=none,file=/dev/...,media=disk,format=raw
  -device nvme-ns,drive=disk0,bus=nvme0,nsid=1
  -drive id=disk1,if=none,file=/dev/...,media=disk,format=raw
  -device nvme-ns,drive=disk1,bus=nvme0,nsid=2

The maximum number of disks (namespaces) is set to 4 since this matches
the maximum number of disks for IDE (although the number for IDE
includes CD-ROMs).

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>