REQ-627 CP-322210 split PCI_plug ops into early/late
For NVidia SRIOV vGPUs the PCI plug operation needs to happen before
QEMU is started. This is possible because these PCI devices are not
advertised to QEMU via QMP and thus don't require QEMU to be running.

This commit splits the list of PCI devices into early (before QEMU) and
(after QEMU) plugs and builds the PCI_plug operation accordingly. The
PCI_plug operation takes a new flag "qmp_add_devive" which indicates
whether the PCI device is advertised to QEMU. Obviously the flag needs
to be false for devices plugged early and true for those plugged late.

This does not create a problem for migration from older releases
because those don't have SRIOV vGPUs and all PCI devices are plugged
late and in the same order as before.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>