CA-327906: don't fail migration if a xenstore directory is missing
Migration of a VM without VIFs got stuck at:
```
Caught Xs_protocol.Enoent("directory"): cleaning up VM stat
```

Some more debugging showed it failed when trying to move this xenstore
entry, which was missing:
```
/xapi/8e8c58cd-eba4-cde3-e1e5-000000000001
```

The intention of this code seems to have been to ignore missing entries,
so do that by ignoring Enoent on the root of a tree.
Do not ignore Enoent on entries deep in the tree since that would
indicate a race condition elsewhere.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>