CP-32446: New hypercall to get MSR_ARCH_CAPS
When MSR leveling is available this will have to get dropped:
at that point all these should become opaque to xenopsd.

For now we have to look at this MSR to make decisions on what CPUIDs
we expose to migrated guests and what CPUIDs we expose to newly started guests.

Signed-off-by: Richard Davies <richard.davies@citrix.com>
Signed-off-by: Edwin Török <edvin.torok@citrix.com>