CA-325940 deactivate NVidia PCI SRIOV b/f using pass-through
An NVidia GPU in SRIOV mode serves a set of PCI VFs. These need to be
deactivated by calling "sriov-manage -d" before the GPU can be
passed-trough in its entirety.  Before binding an NVidia GPU to pciback
(for pass-through), all its virtual functions need to be unbound from
pciback.

When binding the pciback driver, we call new_slot and bind, but so far
only used unbind. This is suspected to cause problems on Intel
architectures and reeks of a resource leak. We now use
unbind_from_pciback() for a proper release.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>