Fix CI: add definitions from new xenctrl.h
The latest Ubuntu 19.10 only has Xen 4.9 headers, so is missing the new
cpuid and msr policy functions, which causes Travis CI and local builds
outside of a container to fail.

Add back the compatibility definitions so that the CI and local builds
work (where we only need linking to succeed and unit tests to run).

Signed-off-by: Edwin Török <edvin.torok@citrix.com>