CA-335206 catch exception from Rpcmarshal.unmarshal
During migration we perform a handshake operation. If it fails, we try
to recover the error by trying to parse the error message as JSON.
However, the message could be something else which leads the JSON parser
to fail. We need to capture this exception in order to generate a proper
error message.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>