CA-332779: verify power state for start/reboot/resume
Do a VM power-state check before executing VM_reboot, VM_start or VM_resume. Do
this in `perform`, right before executing the operation, which should be
exactly the right time. These checks replace similar power-state checks in
xapi, which are done before queueing the operation. This is too early, because
xenopsd may be busy responding to in-guest requests to shutdown or reboot.

The checks are done outside the exception handler that surrounds `perform`,
because we want to just raise an error to the client that requested the
operation (i.e. xapi) without entering `trigger_cleanup_after_failure` (the VM
is good, but the request is bad).

CA-347611: In other to support resuming pre-Lima VMs, it is unfortunately
necessary to allow a Halted power-state for the resume operation rather than
just Suspended. This is because it is impossible to determine whether such VMs
are suspended or not, as the `suspend_memory_bytes` field was not yet in the
persistent VM state.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
(cherry picked from commit 5ce4ea2d6f7d0ca38255af28547032b07eb0d278)