CA-332779: Fix resume unit test due to power-state checks
The power state now matters for the resume unit test. In order to resume a VM, its power state needs to be `Suspended`. This means that the backend internal state for the domain needs to be saved following a suspend.

To make this easier in the high-level unit test, the suspend and resume tests
are combined. Alternatively, VM.export_metadata and VM.import_metadata could be
used to temporarily save the state between suspend/resume tests, but then the
test cases will no longer be independent.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
(cherry picked from commit ada96a7658a6467a01524d7a943e538afab23b64)