CA-344431: ls_lR: add quota
xenopsd watches certain sub-trees in the guest:
vm-data, FIST, drivers, attr, data, control, feature, xenserver/attr

vm-data is replicated as is into the XAPI database, which raises a DoS
concern: each key update is amplified in DB and message-switch traffic:
the entire VM objects needs to be read and written back, which results in
O(N^2) operation if we insert N keys.

VMs do have xenstore quotas, but even when they work they are fairly large
(8192 keys). It is not easy to reduce that quota (calculating number of keys legitimately needed
for 255 disks, 8 NICs, 32 vCPUs, etc.) results in ~7000 keys already.

However vm-data is not meant as a general purpose data store, and although it probably shouldn't have existed in the first place,
we can't change that for LCM releases. Lets at least add a small quota on it, such that N^2 isn't too big.

With this change xenopsd doesn't run out of memory anymore (as long as oxenstored properly enforces quota).
Now a VM should only be able to store ~2.3MiB, instead of ~144MiB which is a significant reduction.

We should consider introducing a limit in bytes, although that would be more complicated because the
 XML serialization can amplify the size (e.g. " gets turned into &quot;) and is a XAPI implementation detail.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Ben Anson <ben.anson@citrix.com>