Add experimental platform:domid VM field
During testing or migration it can be useful to control the domid that a
VM will receive.
By default 0 used to be hardcoded (which meant: let Xen choose).
If `platform:domid` is present then use that. If the domid is already in
use then Xen will give us an error.

Upstream Xen has supported this for a very long time (~15 years),
but it hasn't been exposed in a toolstack until recently.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>