CA-344431: read important xenstore entries first
When the quota is hit some guest agent functionality might stop working.
Warn about this, and try to read the more important entries first
as a best-effort to keep it working (control, etc.).
Read data/ and vm-data/ last, so if the quota is hit there we can still
do a clean shutdown of a VM.

Of course a privileged user inside the guest can still use up entries
beyond the quota in one of the other xenstore subtrees and break guest
agent functionality, but if it is sufficiently privileged it could just
kill the guest agent, so that is expected.

Rename ls-lR-quota to guest-agent-quota.
Add comment explaining depth.

Periodically warn when the quota is exceeded.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
Signed-off-by: Ben Anson <ben.anson@citrix.com>