CA-351685: increase default xenopsd quota
The new byte based quota for 128 entries would be ~1.6MiB,
and for 8192 entries (the previous max before quotas got introduced) ~104MiB.

Dom0 is quite memory constrained, so although 104MiB from a single VM is not enough to OOM it,
if several VMs work together they quite easily can.
However that is when the entries are deliberately written such that they expand to their maximum in xml/json encoding
(e.g. `&` -> `&quot;`).
Normal 8192 entries amount to only ~24 MiB, which is more reasonable.
(8192 max entries * (1024 max key length + 2048 max value length + 10
overhead)).

With the improved fix for XSA-354 in message-switch
having a quota here is less important, the increased memory consumption
would only affect XAPI, and the expanded character sizes should
only happen during saving/loading of the database.
It is prudent to leave it in place to avoid the increased
memory usage by the XAPI database though.

Do not expect this quota to be hit during legitimate usage anymore
though.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>