CA-352061: Call VM.import_metadata directly during localhost migration
In case of a localhost migration, we already have the VM metadata, but
we still export+import to remap VDIs and VIFs. These functions must
now be called directly, because the VM.import_metadata API queues the
operation and blocks (this is not the same on newer branches, where
separate, temporary UUIDs are used for the source and target).

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>