CP-37034: move TSX handling logic out of xenopsd
Reuse the already existing logic in Xen instead of doing this calculation
in the toolstack which becomes outdated on microcode and Xen updates.

Query the HVM/PV Max featureset from Xen and logically OR those feature bits
into `features_hvm` and `features_pv` respectively.

On latest upstream Xen the Max policy tells us what features we can
accept on an incoming migration: i.e. the features are implemented
natively in hardware, emulated by Xen or a compatible workaround by
hardware.
As an example: the TSX instruction may be hidden in the default policy,
i.e. we do not want guests to see it and use it on a fresh boot,
but present in the Max policy.
This is safe because when the host has a TSX instruction the microcode
update would just make the TSX instruction always abort, which is
compliant with the instruction's specification (but then it is not a
very useful instruction anymore, and better hidden on boot).

On Xen 4.13.x there is no Max/default policy split yet, but xenctrl
in the patchqueue provides a Max featureset instead.

This requires a version of Xen that has the new xenctrl stubs.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>
(cherry picked from commit efa6110551e4077017de617e2c4b1d837f0ac0f1)