VM.import_metadata_async: avoid leak
VM.import_metadata_async creates a task and ignores it.
This causes the tasks to accumulate in the tasks and updates list,
causing high CPU and memory usage.

Set a flag on these tasks to automatically destroy them when they finish
(i.e. complete or fail). If the task is already finished at the time we
set the flag (there could be a race), then immediately destroy it.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>