Add platformdata to persistent metadata
VM.create will use the platformdata from the vm record if it is creating
new persistent metadata, when starting the VM. If a persistent record
already exists, in the case of a live migration or resume, it is left
alone as usual.

In the migrate/resume case case, the persistent record will have been
created from stored state by a call to `set_internal_state` when
importing VM metadata. If the incoming state does not contain
platformdata (the upgrade case), this is copied in from the vm record.

When generating the create_info record for Domain.make in
`generate_create_info`, the platformdata is now taken from the
persistent record rather than the vm record.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
(cherry-picked from xen-api/6721700f1ec7b36c5c20a2804661dab006fa9199)