Upgrade featureset in xenopsd when importing VM state
This is to replace what is currently done in xapi. A featureset update
is needed if the VM lands on a host that recognises more features and
therefore has a longer featureset vector. It is done simply by
zero-extending the VM's set to match the length of the host's.

Signed-off-by: Rob Hoes <rob.hoes@citrix.com>
(cherry-picked from xen-api/81454eacfc0efcfe53ace72a7c73234476589ab1)