CP-41796 Close Port 80 (Encrypt Data Transfer During VM Migrations)
This is a backport of several commits; we omit those that in the
original code deal with certificate checking as this is not present in
this code base.

The original code is in xen-api.git but in the backport is split into
xe-api.git, xenopsd.git, and xen-api-libs-transitional.git.

Socket keepalives do not work when stunnel is used, as the given fd is
the local connection to stunnel, and are not set in HTTPS mode. This is
fine, because the stunnel client is already set up with keepalives.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>
Signed-off-by: Christian Lindig <christian.lindig@cloud.com>